feat: add Swagger setup for API documentation and refactor import paths
