+++ Backend/src/swagger.ts
@@ -1,0 +1,41 @@
+import swaggerJSDoc from 'swagger-jsdoc';
+import swaggerUi from 'swagger-ui-express';
+import { Express } from 'express';
+
+const swaggerDefinition = {
+  openapi: '3.0.0',
+  info: {
+    title: 'EasyAssets API',
+    version: '1.0.0',
+    description: 'API documentation for EasyAssets — An Asset Management with QR Codes.',
+  },
+  servers: [
+    {
+      url: 'http://localhost:5000',
+      description: 'Development server',
+    },
+  ],
+  components: {
+    securitySchemes: {
+      sessionCookie: {
+        type: 'apiKey',
+        in: 'cookie',
+        name: 'sessionToken',
+      },
+    },
+  },
+};
+
+const options = {
+  swaggerDefinition,
+  apis: [
+    './src/routes/*.ts',
+    './src/controllers/*.ts',
+  ],
+};
+
+const swaggerSpec = swaggerJSDoc(options);
+
+export const setupSwagger = (app: Express) => {
+  app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
+};
